requirements.txt
# Core dependencies
playwright==1.40.0
asyncio==3.4.3

# Database
sqlite3==3.43.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Web dashboard
streamlit==1.29.0
plotly==5.18.0

# Email notifications
python-dotenv==1.0.0

# Utilities
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Logging
colorlog==6.8.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
